add changes for design oan pdf

+++ app/controllers/users_controller.rb
@@ -9,38 +9,42 @@ class UsersController < ApplicationController
     @auth = env["omniauth.auth"]
     number = (0...6).map { (65 + rand(26)).chr }.join
    begin
-      @user = User.where(provider: @auth.provider, uid: @auth.uid).first_or_initialize do |user|
-        user.provider = @auth.provider
-        user.uid = @auth.uid
-        user.first_name = @auth.info.first_name
-        user.last_name = @auth.info.last_name
-        user.email = @auth.info.email
-        user.current_location = @auth.info.location
-        user.contact_no = @auth.info.phone
-        user.oauth_token = @auth.credentials.token
-        user.oauth_expires_at = ""
-        user.image = open(@auth.info.image) if @auth.info.image.present?
-        user.password = number
-        user.password_confirmation = number 
-        user.role = "candidate"
-        user.save(validate: false)
-        pdf = InvoicingAndReceipts.new("RESUME", @user,@auth)
-        abc = Base64.strict_encode64(pdf.render)
-        decoded_file = Base64.decode64(abc)
-        file = Tempfile.new(["#{user.first_name}",'.pdf'], Rails.root.join('tmp'))
-        file.binmode
-        file.write pdf.render
-        file.close
-        pdf = File.open file
-        pdf_file_name = "#{user.first_name}.pdf"
-        rr = Resume.new(cv: pdf, user_id: user.id)
-        rr.save
-        file.unlink
-        UserMailer.candidate_email_alert(user, nil).deliver_now
-      end
-      if @user.present?
+      @user = User.where(email: @auth.info.email)
+      if !@user.present?
+        @user = User.where(provider: @auth.provider, uid: @auth.uid).first_or_initialize do |user|
+          user.provider = @auth.provider
+          user.uid = @auth.uid
+          user.first_name = @auth.info.first_name
+          user.last_name = @auth.info.last_name
+          user.email = @auth.info.email
+          user.current_location = @auth.info.location
+          user.contact_no = @auth.info.phone
+          user.oauth_token = @auth.credentials.token
+          user.oauth_expires_at = ""
+          user.image = open(@auth.info.image) if @auth.info.image.present?
+          user.password = number
+          user.password_confirmation = number 
+          user.role = "candidate"
+          user.save(validate: false)
+          pdf = InvoicingAndReceipts.new("RESUME", @user,@auth)
+          abc = Base64.strict_encode64(pdf.render)
+          decoded_file = Base64.decode64(abc)
+          file = Tempfile.new(["#{user.first_name}",'.pdf'], Rails.root.join('tmp'))
+          file.binmode
+          file.write pdf.render
+          file.close
+          pdf = File.open file
+          pdf_file_name = "#{user.first_name}.pdf"
+          rr = Resume.new(cv: pdf, user_id: user.id)
+          rr.save
+          file.unlink
+          UserMailer.candidate_email_alert(user, nil).deliver_now
+        end
         sign_in(@user)
         redirect_to root_path, notice: "Welcome! You have signed up Sucessfully and we have email your login Details on your email."
+      elsif @user.present?
+        sign_in(@user.first)
+        redirect_to root_path, notice: "Welcome! You are alredy Register."
       else
         redirect_to root_path, alert: "Sign Up Failed please try again!"
       end
